+++ batch.sh
@@ -31,21 +31,26 @@ IMG="$( ls -1t tmp/*.img | head -1 )"
 ./build-SD-armbian.sh bananapi Bananapi
 
 # Docker x86
-make nextcloudpi-x86 && {
-  docker push ownyourbits/nextcloudpi-x86 
-  docker push ownyourbits/nextcloud-x86 
-  docker push ownyourbits/lamp-x86
-  docker push ownyourbits/debian-ncp-x86
-}
+make nextcloudpi-x86 
+
+# TODO test first
+#&& {
+  #docker push ownyourbits/nextcloudpi-x86 
+  #docker push ownyourbits/nextcloud-x86 
+  #docker push ownyourbits/lamp-x86
+  #docker push ownyourbits/debian-ncp-x86
+#}
 
 # Docker armhf
 cp -n /usr/bin/qemu-arm-static docker-armhf && \
-make nextcloudpi-armhf && {
-  docker push ownyourbits/nextcloudpi-armhf
-  docker push ownyourbits/nextcloud-armhf
-  docker push ownyourbits/lamp-armhf
-  docker push ownyourbits/debian-ncp-armhf
-}
+make nextcloudpi-armhf 
+
+# TODO test first && {
+  #docker push ownyourbits/nextcloudpi-armhf
+  #docker push ownyourbits/nextcloud-armhf
+  #docker push ownyourbits/lamp-armhf
+  #docker push ownyourbits/debian-ncp-armhf
+#}
 rm -f docker-armhf/qemu-arm-static
 
 # License