+++ etc/ncp-config.d/DDNS_freeDNS.sh
@@ -48,7 +48,7 @@ EOF
 
   cd /var/www/nextcloud
   sudo -u www-data php occ config:system:set trusted_domains 3 --value="$DOMAIN_"
-  sudo -u www-data php occ config:system:set overwrite.cli.url --value=https://"$DOMAIN_"
+  sudo -u www-data php occ config:system:set overwrite.cli.url --value=https://"$DOMAIN_"/
 
   echo "FreeDNS client is enabled"
 }