+++ etc/ncp-config.d/DDNS_spDYN.sh
@@ -21,9 +21,6 @@ DESCRIPTION="Free Dynamic DNS provider (need account from spdyn.de)"
 
 install()
 {
-  apt-get update
-  apt-get install -y --no-install-recommends curl # TODO use wget instead
-
   # Create the spdnsUpdater.sh
   mkdir -p "$INSTALLPATH"
   # Write the script to file