+++ etc/ncp-config.d/nc-limits.sh
@@ -32,17 +32,17 @@ configure()
   local CONF=/var/www/nextcloud/.user.ini
   local CURRENT_PHP_MEM="$( grep "^memory_limit" "$CONF" | sed 's|.*=||' )"
   [[ "$MEMORYLIMIT_" == "0" ]] && MEMORYLIMIT_=$AUTOMEM && echo "Using ${AUTOMEM}B for PHP"
-  sed -i "s/post_max_size=.*/post_max_size=$MAXFILESIZE_/"             "$CONF" 
-  sed -i "s/upload_max_filesize=.*/upload_max_filesize=$MAXFILESIZE_/" "$CONF" 
-  sed -i "s/memory_limit=.*/memory_limit=$MEMORYLIMIT_/"               "$CONF" 
+  sed -i "s/^post_max_size=.*/post_max_size=$MAXFILESIZE_/"             "$CONF" 
+  sed -i "s/^upload_max_filesize=.*/upload_max_filesize=$MAXFILESIZE_/" "$CONF" 
+  sed -i "s/^memory_limit=.*/memory_limit=$MEMORYLIMIT_/"               "$CONF" 
 
   # MAX PHP THREADS
   local CONF=/etc/php/${PHPVER}/fpm/pool.d/www.conf
   local CURRENT_THREADS=$( grep "^pm.max_children" "$CONF" | awk '{ print $3 }' )
   [[ "$PHPTHREADS_" == "0" ]] && PHPTHREADS_=$( nproc ) && echo "Using $PHPTHREADS_ PHP threads"
-  sed -i "s|pm.max_children =.*|pm.max_children = $PHPTHREADS_|"           "$CONF"
-  sed -i "s|pm.max_spare_servers =.*|pm.max_spare_servers = $PHPTHREADS_|" "$CONF"
-  sed -i "s|pm.start_servers =.*|pm.start_servers = $PHPTHREADS_|"         "$CONF"
+  sed -i "s|^pm.max_children =.*|pm.max_children = $PHPTHREADS_|"           "$CONF"
+  sed -i "s|^pm.max_spare_servers =.*|pm.max_spare_servers = $PHPTHREADS_|" "$CONF"
+  sed -i "s|^pm.start_servers =.*|pm.start_servers = $PHPTHREADS_|"         "$CONF"
 
   # RESTART PHP
   [[ "$PHPTHREADS_"  != "$CURRENT_THREADS"   ]] || \
@@ -62,7 +62,7 @@ configure()
   local CURRENT_REDIS_MEM=$( grep "^maxmemory" "$CONF" | awk '{ print $2 }' )
   [[ "$REDISMEM_" == "0" ]] && REDISMEM_=$AUTOMEM && echo "Using ${AUTOMEM}B for Redis"
   [[ "$REDISMEM_" != "$CURRENT_REDIS_MEM" ]] && {
-    sed -i "s|maxmemory .*|maxmemory $REDISMEM_|" "$CONF"
+    sed -i "s|^maxmemory .*|maxmemory $REDISMEM_|" "$CONF"
     service redis-server restart
   }
 }